Makefile: don't regenerate yml* targets if RESCAN_DEPS is not FORCE
Targets `images/out` and `build-tools` should be put as order-only-prerequisites:

    https://www.gnu.org/savannah-checkouts/gnu/make/manual/html_node/Prerequisite-Types.html#Prerequisite-Types

Because `images/out` is a folder, which timestamp is always changes if new
files are modified inside and `build-tools` is .PHONY target.

This patch moves `images/out` and `build-tools` to the order-only-prerequisites
to avoid constant yml* regeneration, which forces rebuild images.

The force of rebuilding is controlled by the RESCAN_DEPS variable, which is
set to FORCE (is .PHONY target), which means the default behavior won't break,
but gives us freedom to reset the RESCAN_DEPS and do not generate live image
if it was already generated.

This 'do-not-generate' live image logic will be used in the next patch for
the 'LIVE_PKG' update.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>