Prefer physical SIM slot over eSIM
If the user does not select a specific SIM slot, EVE will utilize
whichever SIM slot is already designated as primary (e.g., by the device
manufacturer).
However, there is one exception: if the primary SIM is an eSIM
(virtual embedded SIM), EVE will default to using the first physical
slot instead, as eSIM is not supported. EVE will attempt to connect
(and fail) with the eSIM only if the user explicitly selects the eSIM
slot.

Signed-off-by: Milan Lenco <milan@zededa.com>