collect-info.sh: output log lines instead of JSON
This changes default behavior of the collect-info.sh when invoked
in the collected folder:

  $ ./collect-info.sh
  2024-03-11 11:26:23.078|debug|dhcpcd|| eth0: renewing lease of 192.168.128.123
  2024-03-11 11:26:23.078|debug|dhcpcd|| eth0: sending REQUEST (xid 0x65463210), next in 3.5 seconds
  2024-03-11 11:26:23.080|debug|dhcpcd|| eth0: acknowledged 192.168.128.123 from 192.168.128.1
  2024-03-11 11:26:23.080|debug|dhcpcd|| eth0: leased 192.168.128.123 for 600 seconds
  ...

so by default outputs human readable log lines in the format:

   DATETIME|SEVERITY|SOURCE|FILE|CONTENT

for the kernel or system output the FILE can be missing.

If you need the output in json provide the '-j' option, so

  $ ./collect-info.sh -j

This will output every message wrapped in JSON, as it was before this commit.

Be aware, that EVE has nested JSON, which the collect-info tries to unwrap
as well with the following line:

  (.content as $cont | $cont | try (fromjson.msg) catch $cont)

So that if nested JSON is valid we get ".msg" field from it, if not valid,
then output the whole ".content".

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>