grub: Add device tree loading for Siemens IPC 520A
If the Siemens IPC 520A device is detected, load the corresponding
device tree. The detection is based on the asset TAG value because the
factory BIOS flashed to the QSPI memory is the default one that comes
with JetPack, so the device is also listed as a NVIDIA Jetson Xavier NX
Developer Kit by SMBIOS.

The installation behavior is the same as for Lenovo SE70 device, i.e.,
EVE it's installed in the NVMe SSD and the eMMC it's wiped off.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>