pkg/kdump: slow down printing the dmesg of the crashed kernel
Print dmesg of the crashed kernel line by line with some reasonable
sleep inbetween lines to slow down the output on the console and give
a chance to record the stack properly. Typical use case is BMC which
records with very low frame rate so the video contains the bottom of
the long stack only but the top is missing. This happens if the stack
is huge and occupies more than 1 screen height.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>