PubSub: introduce memdriver
In order to test that service is working starting from publications
and subscriptions memdriver was introduced.

This structure allows developer to create PubSub, send publications,
so that tested component could recieve them. There is no need to
expose publications and subscriptions as exported fields in service
structure so that you could configure them like it's done in
parseconfig_test.go.

Note that only persistent subscription will be updated due to implementation

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>