Do not assume that SIM is absent if its type is not known
I have seen cases where modem would not report SIM card type
(the returned value was not specified), even though the (physical)
SIM card was inserted.
If SIM card type is not known, we can instead determine the card
presence based on the ICCID number availability.

Moreover, when ModemManager is unable to recognize the SIM card type,
EVE would report the type as SimTypePhysical. This is clearly
incorrect, SimTypeUnspecified should be reported instead.

Signed-off-by: Milan Lenco <milan@zededa.com>