Allow device access when loading OCI spec from file
Since the update to runc v1.1.0 and containerd v1.6.1 in
158cecdb9a021a403461f466b1eb0a72ffc009cf we need to explicitly allow
device access in the OCI config. This is done when generating a new OCI
spec.
However, when loading an OCI config from a file, we need to make sure
that the device access is added, because older OCI configs may lack
this configuration.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>