pillar: Release CPUs on domain activation failure.
This commit addresses an issue where CPUs assigned to a domain within
doActivate() are not released if the domain activation fails. The new
logic ensures that CPUs are properly released and the CPU mask in the
status is updated accordingly. This is achieved by introducing the
releaseCPUs function and calling it in the appropriate error handling
blocks within doActivate().

It is common for doActivate to fail in scenarios such as switching
application profiles that share the same adapter. In such cases, the
second application will fail to activate until the first one releases
the necessary adapter.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>
(cherry picked from commit 10c924ee953da4a5f1233bd22acf0bec02e85566)