Incorporate Longhorn 1.6.2 fixes
Longhorn 1.6.2 includes a number of improvements
including an apiserver performance improvement
and an rpc deadlock fix

This commit has to drop the CDI version to handle
competing security changes in CDI and longhorn as
the default configuration of Longhorn 1.6.1+
is incompatible with CDI 1.55+.  This needs to be
resolved but in the short term the updated longhorn
is needed more than the updated CDI, the CDI 1.57.0
release was primarily included for a qcow2 performance
improvement on slower devices.

The changes which broke compatibility appear to be:
CDI 1.55+ started running uploads as a nonroot kvm user.
Longhorn 1.6.1+ started adding volume devices to a disk group.
There is a configuration change available to resolve this issue
but analysis of the changes to security are ongoing.

Fix NewPVCDefinition default AccessMode to RWO.  An earlier merge mistakenly
changed it to RWX-Block for non downloaded volumes which is a
combination not supported by Longhorn.

Signed-off-by: Andrew Durbin <andrewd@zededa.com>