pillar: introduce `debug.enable.vnc.shim.vm` per-node flag
For security reasons VNC access to the shim VM is disabled by default.
To enable such access the `debug.enable.vnc.shim.vm` flag has to be
set to true. The flag is global per node.

This flag complements the original `EnableVncShimVm` flag introduced
in this commit:

    7545e4df61e4 ("Add the EnableVncShimVm flag")

which acts per application and requires controller assistance.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>