memory-monitor: Introduce memory monitor tool.
The tool monitors memory usage of the zedbox process, the eve and pillar
cgroups and triggers a handler script when certain thresholds are
reached or a memory pressure event happens. The monitoring system is
implemented in C and includes a configuration file for customizing the
thresholds.

Additionally, this commit includes a pressure test tool that allocates a
specified amount of memory to simulate memory pressure.

For more detailed information about the memory monitoring system, its
configuration, and how to use the pressure test tool, please refer to
the README file.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>