memory-monitor: Include sys/select.h explicitly.
The `sys/select.h` header file is necessary in the case of musl library
because it provides the definitions for the `select()` function and the
`fd_set` data type. The musl library, which is designed to be
lightweight and simple, does not include these definitions indirectly
through other header files. Therefore, when using musl, we need to
include `sys/select.h` directly.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>