memory-monitor: Add memory-monitor config update command to eve.
Introduce a new command to the EVE CLI, memory-monitor-update-config,
which allows for the dynamic updating of the memory monitor's
configuration. This enhancement provides a more efficient way to update
the memory monitor's configuration without having to manually send a
SIGHUP signal to the memory monitor process.

The README documentation has also been updated to reflect this new
feature, providing users with additional guidance on how to use the new
command.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>