Implement configurable device port MTU.
User is now able to set the MTU inside the Network object assigned to
device network port (aka SystemAdapter).
By default (if not set), EVE will set the default MTU size, which
depends on the network adapter type. Ethernet and WiFi adapters
default to 1500 bytes, while cellular modems typically receive
their MTU value from the network provider, which EVE will use
unless the user overrides the MTU value.

For Ethernet and WiFi ports, MTU is applied by NIM microservice.
For cellular modems, MTU is managed by the wwan microservice.

As part of this commit, few changes in the area of config items
of DPC Reconciler were done. For example, PhysIf and IOHandle
were refactored into:
- NetIO: represents existence of a NIC (previously the purpose of PhysIf)
- PhysIf: replacing IOHandle and additionally managing the network
  interface created for the NIC by the kernel (currently only the MTU
  attribute is managed)

Please note that this commit does not include MTU support for network
instances and applications. That part of the MTU support was split and
will be delivered as a separate commit.

Signed-off-by: Milan Lenco <milan@zededa.com>