pillar/zedagent: introduce `ForceSend` flag for distanation bitset
The flag acts as a hint for request handlers and tells that
request should be sent in any case. In the next patches the
NTP sources reports will be added and will be sent to the
controller if NTP peers have been changed, so a comparison
will take place. LOC destination is special and should
bypass any checks and request should be forcibly sent.

This scenario is the main user of the `ForceSend` flag.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>