Manually assign uid/gid to username and groupname in dom0
Linuxkit dosn't use container's /etc/passwd and /etc/group files to
resolve user and group names defined in build.yml. Instead it assignes
increamentally created ids to the container[1] based on their delared
position in rootfs.yml. It only respects the container's build.yml
uid/gid value if its integer[2].

By assigning a fixed uid/gid in dom0, we can use the same value in the
build.yml of the containers and be sure that access to resources work as
expected, and adding/reordering containers in rootfs.yml won't break
the access control.

[1] https://github.com/linuxkit/linuxkit/blob/4f89f4f67e392ffa8c8bab63dfaf31746e3c11a0/src/cmd/linuxkit/moby/build/build.go#L188
[2] https://github.com/linuxkit/linuxkit/blob/4f89f4f67e392ffa8c8bab63dfaf31746e3c11a0/src/cmd/linuxkit/moby/config.go#L708

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>