memory-monitor: Preserve errno in SIGHUP handler.
Ensure the value of errno is preserved during the execution of the
SIGHUP signal handler.

Signal handlers can interrupt the normal flow of a program and
potentially change errno. This is because signal handlers run in the
context of the thread that was interrupted by the signal. If a signal
occurs while the thread is in a system call, the signal handler might
change errno.

By storing the errno value at the start of the handler and restoring it
before the handler returns, we maintain the correctness of error
handling in the program. This practice is particularly important in
non-terminating signal handlers where the program execution continues
after the handler.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>