pillar/device-steps.sh: reload NTP sources instead of NTP daemon restart
When NTP sources are updated, chronyd needs to be notified. The easiest
way was to kill the chronyd instance and restart it with new NTP sources.
This restart seems is not very much appreciated by the watchdog, which
detects absense of the process by the pidfile and fires (see the #4014)

I could not reproduce the issue locally, but I see the following logs
quite often:

   2024-06-28T13:23:28.680584031Z;watchdog;watchdog: pidfile: /run/chronyd.pid
   2024-06-28T13:23:43.234635519Z;pillar.out;2024-06-28T13:23:43,207263478+00:00 NTP (chronyd) server 3155 still running
   2024-06-28T13:23:43.774360721Z;watchdog;watchdog: cannot open /run/chronyd.pid (errno = 2 = 'No such file or directory')
   2024-06-28T13:23:44.77729473Z;watchdog;watchdog: cannot open /run/chronyd.pid (errno = 2 = 'No such file or directory')
   2024-06-28T13:23:45.77819751Z;watchdog;watchdog: cannot open /run/chronyd.pid (errno = 2 = 'No such file or directory')

which happen during the chronyd restart. I do not have an answer why
restart could happen so long, but anyway this is not a good approach
to restart a daemon for new NTP sources population.

This patch deletes old sources and populates new list. Unfortunately
chronyc does not provide a way to delete all sources in one line,
so this has to be implemented in a loop.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>