Wait for wwan to apply latest config before testing connectivity
When cellular network config changes, it may take few seconds for the
wwan microservice and the modem to apply the new config.
NIM microservice should wait for the wwan config to be fully applied
when cellular connectivity is used for management and ethernet is not
available. Otherwise, it may determine the connectivity as not working
and failover to the previous config too fast, not giving the modem the
chance to connect with the latest config.

Also fixed in this commit is the Equal() method for CellularAccessPoint.
It used to compare also timestamp inside ErrorAndTime which may change
even if there is no actual config change. As a result, wwan microservice
would trigger re-connection even when there was no config change to
apply. This fix is in pillar and will be propagated into wwan/mmagent
in the next commit (with "go get") once this one is merged.

Signed-off-by: Milan Lenco <milan@zededa.com>