Login to dockerhub before building eve
When building eve and pulling container dependencies, we may hit the
Docker Hub imposed limit of 100 pulls per day (for anonymous pulls).
Step to solving this problem is to login into Docker Hub before pulling,
That way we will get 200 pulls per day instead of 100 (which are
potentially shared with other users of BuildJet runners).
We can also consider upgrading our dockerhub account to the Pro version
with 5000 pulls a day.

Signed-off-by: Milan Lenco <milan@zededa.com>