pillar: scripts: Do not remove directory for dhcpcd
The dhcpcd.sh script creates the /run/task/vifs/<APP_UUID>/ directories on
an up command (configured as a Prestart hook in the container OCI
interface) and removes it entirely on a down command (configured as
Poststop). However, the etc/resolv.conf.new file is created by pillar and
should be mounted inside the container. If any issue arises while setting
up the bridge + virtual network interface during container initialization,
pillar will retry to start the container, but at this point the
etc/resolv.conf.new will not be available anymore since it was removed by
along with the other directories created by the dhcpcd.sh script.

This commit solves this issue by simply not removing the entire
directory, but only the resolv.conf file that is created during the
setup. Pillar already handles the creation and removal of this
directory, so no changes are required on his side and the directory will
be removed when not needed anymore.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>