pillar: Remove full device access to native containers
Access to all file devices are granted to QEMU container loader in order
to be able to access kvm and other I/O devices. The access is included
during the setup of the OCI spec for the container in
pkg/pillar/containerd/oci.go. However, native deployed containers will
also have access to all file devices from the host. This commit moves
the corresponding code to the hypervisor setup functions, so native
deployed containers will not have full access to the host.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>