pillar: Add support to CDI for native containers
The Container Device Interface (CDI) is already available on containerd and
that are CDI files provided for NVIDIA platform that describes GPU devices
for native container access. This commit introduces the support for CDI on
pillar, so Video I/O adapters can pass a CDI device name to enable the
direct access on native containers.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>