Allow attaching volumes to multiple app instances
Instead of having single VolumeRefConfig and VolumeRefStatus per
volume, we'll have a single VolumeRefConfig/Status pair per "link",
connecting app instance to an attached volume,
where a link is a tuple of volume unique ID and app instance ID.

Further we simplify the volume reference counting by counting the number
of known reference objects instead of summing the reference counts. This
allows us to get rid of the reference count field in VolumeConfig,
VolumeRefCount and VolumeRefStatus, only keeping it in VolumeStatus.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>