handle empty image-config.json in containers
If image-config.json is empty, it indicates an error in building the
JSON OCI spec of the image. This commit makes sure we sync the file
and the directory after writing the OCI spec.
Additionally, when reading the file from disk, it checks the presence of
the containers rootfs directory.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>