Make flow-logging configurable
Flow-logging is an EVE feature that enables the collection of information
about every network flow in an application. This includes attributes such
as source/destination IP, source/destination port, applied ACL rules,
packet/byte counters, and more.

However, this feature is always enabled for every network instance, which
can generate a significant amount of data that is subsequently uploaded
to the cloud. This can place a burden on both the network and the controller
database.

Another drawback of flow-logging is that the iptables rules that EVE
installs for network instances become significantly more complex due to
this feature, introducing additional packet processing overhead.
Specifically, the mangle table is used to mark every connection, routing
it either towards its destination or into the dummy blackhole interface.

This commit introduces a configurable flow-logging option. When disabled,
EVE no longer marks connections, the iptables mangle table is not used,
the blackhole interface and associated configuration items are not created,
and most importantly, the flow collection routine is deactivated.

Please note that the solution for prohibiting traffic forwarding between
device ports (see commit 860f4d63f9c071d6ef9605364b5bd5a7c984824b)
had to be revised. The previous solution relied on connection marking.
The new, considerably simpler solution is based on the filter table and
functions the same way, regardless of whether flow-logging is enabled
or disabled.

Signed-off-by: Milan Lenco <milan@zededa.com>