fix: use tag to set rootfs version only on tag
The rootfs version should be set to tag only if building from a clean
tagged state. Otherwise we have a different version naming scheme.

This commit fixes the `git describe --tags` command to match the tag
exactly, because it default behavior is to "suffixes the tag name with
the number of additional commits on top of the tagged object and the
abbreviated object name of the most recent commit". So a simple match
would always return the tag related rootfs version like
12.6.0-2-g8d558024e.

Also this commit ensures that each part of the version (major, minor,
patch) has at least one digit, which is a more accurate representation
of semantic versioning.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>