Avoid overwriting the original error from the failed app network creation.
Retrying a failed attempt to connect an application to a network instance
mistakenly repeats the call to doActivateAppNetwork instead of calling
doUpdateActivatedAppNetwork (if it failed after/during activation).
The NIReconciler does not accept duplicate calls to AddAppConn and will
return an error:
doActivateAppNetwork: failed to activate application network: NI Reconciler: App <uuid> is already connected.

Zedrouter should instead call UpdateAppConn in the attempt to fix the
config.

Worse, this error from the retry attempt overwrites the original error,
potentially misleading the user and obscuring the real root cause of the
failed app network creation.

Signed-off-by: Milan Lenco <milan@zededa.com>