Introducing portprober as successor to uplinkprober
As we generalize the concept of NI port probing, which is currently
limited to uplink & freeuplink shared labels and one port selected
for all routes, into a multipath routing with probing-based port
selection with user-defined shared labels, it is necessary to replace
uplinkprober with a more advanced component
(and preferrably avoid using "uplink" for the name).
See: https://github.com/lf-edge/eve-api/pull/53

This commit introduces portprober - a sucessor to uplinkprober.
It will be used by zedrouter to probe port connectivity for multipath
IP routes. These are routes that select multiple possible output ports
using a shared label. This can be EVE-defined label, such as "uplink"
matching all mgmt ports, or a user-defined label selecting a custom
subset of ports.
For now we will not support load-balancing and zedrouter will have to
pick one output port at a time for each multipath route. This will be
done by portprober. It will probe connectivity of each port (with
possibly user-defined probing endpoint and probing method) and based
on the results and some other criteria such as cost, wwan signal
strength, etc., it will pick the best port. The probing algorithm is
pretty much the same as implemented in uplinkprober, just extended to
support user-defined shared labels, user-defined probing method, etc.

Note that in this commit we just add the portprober package and make
only minimal changes in the pillar/types package (e.g. introducing
SharedLabels into DPC & DNS).
In the follow-up commit(s), uplinkprober will be removed and zedrouter
will be updated to support multipath routes and to use portprober
instead.

Signed-off-by: Milan Lenco <milan@zededa.com>