Do not mix up IPv4 and IPv6 in NAT iptables
Recent enhancements done in NI Reconciler introduced a bug in the config
prepared for NAT ip(6)tables, where we may accidentally attempt to create
a rule that mixes IPv4 and IPv6 addresses.

Note that I have also added some IP version checks in places which are
currently not reachable because we do not support IPv6 Local network
instance. Just in case.

A small renaming ipv6 -> forIPv6 was also added for more clarity.

Signed-off-by: Milan Lenco <milan@zededa.com>