eve-tool: block until "verbose off" can be run
When eve boots and one runs "eve verbose off" it often
does not work because the process it sends the SIGSTOP to
does not exist yet.
Instead, let's wait until this process exists and then send
the SIGSTOP.
This way the user does not have to run
sh -x /bin/eve verbose off
in a loop and check if the PID has been found.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>