memory-monitor: Add memory pressure statistics collector.
Implement a daemon to collect memory pressure statistics from
`/proc/pressure/memory`. The collector reads data every second and
writes it to an output file. The output file includes a header and the
columns for all the data fields available in the original proc file.
Ensure data is written to the file immediately using fsync to avoid data
loss in case of a crash. Daemonize the process to run in the
background.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>