memory-monitor: Update README with PSI statistics collector instructions.
This commit updates the README file to include detailed instructions for
using the new PSI statistics collector tool.  The README now includes
instructions on how to build the tool using the Makefile, deploy it to
the EVE system, and run it. Additionally, the README provides guidance
on how to analyze the collected PSI statistics using the visualization
tool located in the statistics-collector directory.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>