add activate credential service to metadata server
Activate credential provides proof that the Endorsement Key (EK) and
a signig key (in this case eve created AIK) are owned by the same TPM.
This is way to extend the trust from EK (which theoretically comes with
OEM certificate) to a arbitrary TPM resident, restricted, signing key.

The added service allows to stablish trust to the AIK and sign arbitrary
data with it.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>