dom0-ztools/eve: Add commands for PSI collector control in the eve script.
This commit extends the `eve` script with new commands `psi-collector
start|stop`. These commands interact with the internal debug API to
start and stop the Memory PSI collector, respectively.

This addition makes it convenient to control the PSI collector directly
from the terminal, with the results stored in
`/persist/memory-monitor/output/psi.txt`.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>