add PROGRAM parameter to eve-enter-container
The script eve-enter-container calls internally nsenter, which Defaults
to calling $(SHELL) if no PROGRAM parameter is provided. And since the
shim VM is built from alpine image, the default shell is /bin/ash.
So unless the container running in the shim VM is also based on alpine,
it won't have /bin/ash and the script will fail.

This commit adds a PROGRAM parameter to the script, so that the user can
specify the shell (or another program) to be used when entering the
container. If the parameter is not provided, the default is /bin/sh.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>