add tty mux functionality to chroot2
This commit adds the ability to control the container's entrypoint
process from multiple tty consoles. Mainly it's done to support the
simultanious access to the container's console from the serial terminal
and the VNC console.

The `chroot2` script now accepts two additional argument which are the
names of the tty consoles (serial and VNC) to be multiplexed. The script
accepts inputs from both consoles and forwards them to the container's
entrypoint stdin. The output from the container's entrypoint stdout and
stderr is also multiplexed to both consoles. Additionally both consoles
forward the common signals such as Ctrl+C, Ctrl+Z, Ctrl+\, Ctrl+D to the
container's entrypoint process.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>