github: Add GH Action to check SPDX license headers in new files.
Created a workflow to automatically check for SPDX license headers in
new files when a PR is opened or changes are pushed to `master` or
version-stable branches.

Added `tools/spdx-check.sh` script that scans new files with specific
extensions and certain filenames for the `SPDX-License-Identifier`
header.

The script excludes files in `vendor/` directories from the check.

The workflow will fail if any of the new files are missing the required
SPDX license header.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>