Print messages on /dev/tty instead of /dev/console
/dev/console ins the current console for kernel so even when we switch
to /dev/tty3 messages still appear and corrupt monitor application. Fo
now print them on /dev/tty which is the default kernel tty.

TODO:
1. diag output must appear on /dev/ttyS* if defined in the list of
   kernel consoles
2. The same applies to other subsystems

In order to do so the kernel command line must be parsed and the last
console= must be detected. Unfortunately there is no easier way to do
this

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>