vcomlink : add the vcomlink agent to the pillar
This commit adds the vcomlink agent to the pillar. The vcomlink agent
is a new agent that will be used for communication between the VM and
the host. The vcomlink agent is used to server commands from the VM
(at the moment only TPM related commands), the idea is to expand it to
support more commands like VM CPU/Mem reporting, security queries,
networking queries, etc.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>